Fix parsing of XTCE container sheet when no offset to parent is specified
If no offset to the parent is specified all sequence entries of that
container should be relative to the previous entry. With the previous
logic only the first sequence entry would be relative, the
absoluteoffset variable would then reset to a low value leading to
wrong results from the second entry onwards (which would be wrongly
considered as being of type CONTAINER_START).