Support searching for an Opsname
When pressing <CTRL-F>, a Find-Bar now pops up in the parameters table.
The bar can be easily closed with <ESC>. All search actions are also
accessible through buttons on the find-bar.

At the moment you still need to press ENTER before it searches.
There's no instantaneous highlighting as of yet.

Note that this commit also includes some changes that remove all kinds
of unneeded borders on the default Swing components.