Replace textual enum desc with an lov
The enum description in the calibration column of the parameters
table is often very long. It makes it very hard for the user to
see all the possible values an enum can take.

This commit replaces that description with a link. Upon clicking
the link, a screen pops up with a table with all possible values
the enum can take. The actual value is pre-selected.