Only output FileNotFoundEx when serialized db exists
When adding a new instance to the mdb.yaml file, the first time
afterwards when launching yamcs, a nasty stacktrace will be printed to
the console. The reason being that a file *.consistency_date could not
be found. This check should not happen if the DB has never been
serialized before (i.e. the first time).