Fix processing of dataencoding for 3-byte integers
There was a unit test for 3-byte integers, where the number 99
(0000_0110_0011)b was successfully decoded. Changing the number 99
with for example the number 236 (0000_1110_1100)b would make the unit
test fail. The reason being that the 3-byte integer is read in two
goes (first a short, then a byte). Both of these are read as signed,
whereas we want to be working with the unsigned version.