Slightly modify structure of replay requests
This was done to make the API of packet retrieval consistent
with that of parameters, and also to make it possible to add
future conditions to any replay type.

Previous API in proto file is marked deprecated. Unfortunately
protoc 2.3.0 does not generate @Deprecated annotations. Code comments
were added where multiple API versions are still supported.

Also in this commit:
 - the CliPacketDump was fixed. It did not work due to some small bugs
 - while the .proto file stated that requesting all containers is
   possible by not specifying a filter, in practice it did not seem to
   work. So this got fixed as well (taking privileges into account).