Preserve whitespace in USS labels
This commit also changes the default font-family to
monospace. Lucida Sans Typewriter is a monospaced
font as well.

We encountered some case where a USS display depends
on the fixed-width property of the font, in order to
align two labels using spaces.

The combination of these two changes supports this.