Support user-defined event algorithms
This commit adds event functionality to the Yamcs ScriptHelper
available from a CustomAlgorithm context.

By calling for example Yamcs.info("msg"); from within the
JavaScript, an Event "msg" will be sent to the EventViewer.

The DerivedValues_XTCE class was modified so that it would
support algorithms without any output parameters. This is still
compatible with XTCE.