Introduce XTCE Algorithms
All the sliding-window stuff and event generation that was
introduced in recent commits was mostly reverted and replaced
with the new notion of Algorithms.

This provides more flexibility, and is more compliant to XTCE
than derived values were.

Algorithms are defined in a new tab in the XTCE spreadsheet
(XML support forthcoming). Some new features:
 - Support for referring to past values
 - parameters can come from any spacesystem in the hierarchy
 - multiple out parameters are possible