Support configuration of custom parameter providers
In the yaml file it is now possible to define the list
(and order) of parameter providers to be loaded by the
ParameterRequestManager.  The only exception is the
processed-parameter provider for which no solution has
been found yet (it's still added by the Channel).

Note that the results of the AlgorithmManager are now
no longer included in the same delivery as the required
input arguments. The algorithm results pass again
through the ParameterRequestManager, resulting in a new
delivery. A side-effect of this, is that we now can have
algorithms that depend on the out of another algorithm.

Care should be taken though, as no checks for circular
dependencies are in place.

In order to maintain backwards compatible, the
ParameterRequestManager will still revert to previous
behaviour when the user does not define any providers
in the yaml file.

Note that support has been dropped for the DerivedValuesManager
Adding it as a parameter provider will not cause any
exceptions, it just won't have any effect.