Allow reading custom tabs from XTCE spreadsheet
Opens up SpreadsheetLoader so that it can be extended.
The new method loadNonStandardSheets() is a hook that
is called just before loading algorithms.

Custom extensions can load any kind of data (not necessarily
parameters), as long as it is grouped in a NonStandardData
implementation. The NonStandardData interface requires
implementation of two operations:
 - Applying the fully qualified name (could have an impact
   on the aliases of any non-standard parameters)
 - How the non-standard data needs to be combined with
   non-standard data of the same type, coming from a sub
   space system.

These are the same operations that are implemented in the
conversion of a hierarchy of space systems to one XtceDb.

The mdb.yaml already provides enough functionality to enable
the custom SpreadsheetLoader for the chosen space system.