EventViewer: fix bug when event msg is longer than 10 lines
The detailed view of an event shows all the text areas collapsed
when the event message is longer than 10 lines. The cause seemed
to be the weightx and weighty properties of the msg text area's
GBC.

This commit also fixes another small bug where the text area
would by default be always scrolled down to the bottom.