Expose Yamcs archive through a web api
A first go at a web api, it uses chunked transfer because we
don't want to store to a temp file just because of the
content-length. Chunks can be pretty small, they are aligned
with the incoming data from the ReplayServer.

Only parameters & packets are supported, and profiles are
supposed to be predefined in <cache-dir>/profiles/