Fix packetviewer crashes when switching versions
The interpretation of the UI prefs for 'RecentFiles'
is a Vector in old Yamcs versions, and an ArrayList
in new Yamcs versions. This causes class cast
exceptions when switching between versions.

Clients with this commit integrated, will from
now on store this pref under 'RecentlyOpened'.

From that point on, old clients will no longer
crash. It might be necessary to manually
remove the pref one more time though.