Save column order of EventViewer in UI prefs
Save the column of an order with respect to the other columns in
the user preferences when the user closes the window.

When something in the configuration has changed (e.g. extra columns
are no longer present), it will do its best to restore the default
Event Viewer look.

This commit also brings a couple of changes to borders and colors
that appeared while testing the Event Viewer on Mac.