Algorithms: Support libraries and other languages
This adds some configuration options to an AlgorithmManager, where the
algorithms of an instance can be switched to something other than JavaScript.
For example when including jython-standalone on the classpath, all algorithms
could be written in Python.

Additionally, it moves the earlier introduced configuration option for
specifying libraries, from mdb.yaml to yamcs.<inst>.yaml, which seems like
a better fit.