Make bitsize optional for (most) output parameters of algorithms
Outputs of algorithms don’t really need raw bit sizes. This commit
makes it optional (it’s still allowed, which could be exceptionally
useful).

In case a calibration is shared with another parameter, the bitsize
should still be specified to be the same as that other parameter.

There will now be an error message if this is not the case.

In the future it would be helpful if there is an exception to this for
output parameters, but that’s a bit of a hassle right now.