Add XTCE-like alarms support for integers, floats & enums
The spreadsheet structure is a little bit different than the
previous Limits sheet. Low/high columns were removed, and
replaced by a trigger type/ trigger value column. Low/high are
now two different trigger types. An additional trigger type
‘state’ was added which allows triggering on the state of
Enumerated parameters. There appeared to be already
some logic in place for this, but not through the
spreadsheetloader.

Severity levels were extended. This was also
reflected in an update of MonitoringResult, which
was not using the XTCE alarm levels.