Add minViolations attribute support
XTCE has a minViolations attribute per AlarmType. It’s a bit obscure how
exactly it should be interpreted. The implemented approach is
to keep setting the appropriate pval MonitoringResult
(e.g. WARNING_LOW), even if minViolations is not reached. However
the minViolations is used to limit when an alarm event is sent.