Improve error handling in PV (fixes #3)
In order not to pollute to much code,
DatabaseLoadException was made unchecked.
I consider this a better fit anyway. PV will still
check it and render it in a nice dialog.

The old XtceDbFactory was not very friendly towards
its users. It would forcefully exit which was causing
the PV to disappear on the slightest user error in the
spreadsheet.