Prepare algo engine scopes for concurrent execution
So far we could get away with just using one ScriptEngine, and
reusing it for all algorithms. However, scriptengines are not
guaranteed to be threadsafe, and sometimes unwanted effects
can happen (e.g. this is the case with the current algorithm name
which is set on the engine before any run.)

Furthermore the current code did not make variables declared
in a library global (as in ‘shared’) for all algorithms. It now does
it correctly.