Unify PP and XTCEDB spreadsheet loading
Unifies the logic of the PP and XTCEDB spreadsheet loading.
This makes it possible to load PPs from the same sheet as
where the other sheets are defined.

A separate PP loader is still maintained for cases where
the separation is preferred, but it now redirects all loading
to the unified SpreadsheetLoader.

Another advantage is that custom Spreadsheetloaders can
now choose which sheets they want to load and in which
order. This opens the possibility of e.g. supporting PP and
calibrations at the same time.