AB: Stretch content when window is resized (fixes #1)
When the zoomstack is at level 1 (meaning user has not
zoomed in yet), resizing the window will stretch the content
(showing same data). Kind of defeats the purpose of the
BufferedImage, but hey. Maybe something to improve
in the future.

When the zoomstack is above level 1 (meaning user has zoomed
at least once already), the current behaviour remains unchanged.
We may update this in the future as well, but this is a more difficult
(and less likely) occurence. There's also usability concerns, in that,
if you've stretched the content, you can't return back to your original
view when you resize the window smaller again.
Filling in with a hatched bg would be my pref for this case.