Add commanding schema for rest api
This also renames the proto packages of all proto-files. The change was
done in a a way that does not influence the binary format, nor the api.

The reason for doing this, is that protoc does not support multiple
enum states in different enums that are named the same. This restriction
only holds within a proto package. Since all packages were named
org.yamcs.protobuf, this occurred when launching protoc on the rest.api
which includes various other proto-files.
The state DISABLED happened to be present in both MonitoringResult and
QueueState.

The solution was between this or prefixing the QueueState enum states
which would break the API exposed by yamcs-api.

See https://code.google.com/p/protobuf/issues/detail?id=515