Increase logging level of all sent rest exceptions
Just to clarify positions here: when rest exceptions are sent. Only
internal server errors should get a priority treatment in the yamcs
server log. All other exceptions are client-related and should only
be visible in the server with a finer logging level.

This commit just increases the previous trace to debug, which makes
it a bit easier for developers to debug rest exceptions for any
operation.