Send updates on processors/clients over web socket
This provides similar behaviour to what was previously available only
through hornetq. whereas hornetq works with an 'eventName' field to
distinguish when two messages of the same type are incoming, the web
socket api, will instead send an extra state-flag in the message itself.

The committed example of this, is ClientInfo which is returned on both
registerClient, and unregisterClient. The user of the websocket-api can
check the new 'state'-flag to see whether the provided ClientInfo to
distinguish between CONNECTED / DISCONNECTED.

For ProcessorInfo, there was already a ServiceState-flag which seems
sufficient.