Send current set of processors/clients upon websocket mgmt subscribe
This makes a change in the YProcessor that should be reviewed.
The CopyOnWriteList (which was previously only used by the HornetQ
management stuff), will now contain also entries for listeners coming
from the websocket-api.

In order to prevent duplicates (similar to other management-updates),
the collection was changed to an CopyOnWriteSet. Any iterators that were
using indexed access, therefore, had to be replaced. They now work
with a for-each, which AFAIK is fine with the CopyOnWriteSet. (i.e.
no excessive overhead).