Make TagDb a bit friendlier for requesters that prefer IDs
The upsert and delete operation from the TagDb relied on an
incoming ArchiveTag argument specifying an existing tag in
the TC table. When working through the REST API this is not
ideal, as we would prefer to pass just the ID as part of the
request uri for example when doing an update.

This commit sort-of solves that, although the fact that the id
is not really an id, but needs to be considered together with
the start-time is still slightly annoying.