Refactor REST API, introduce RestRequest and RestResponse
This is a first attempt at reducing the number of arguments that
are passed during the routing of an incoming request.

The logic for implementors is pretty simple. All incoming request
variables are encapsulated in a RestRequest and every method should
return either a RestResponse, or throw up a RestException.

A next refactoring might deal with exceptions a bit differently, they
are currently still in AbstractRestRequestHandler, but for various
reasons it might make more sense as sth on par of RestResponse.