Fix corner-case where REST handler was not returning 500 exceptions
If an unexpected error happend, like for example a NullPointerException.
The server would sometimes not return an error 500 to the client, which
meant that the connection was left open.

The reason was due to a cascaded NPE in the protobuf building phase when
the message of the caught exception was 'null'.