CompactFormatter: Prevent concatting throwable to error message
When using a log-method and providing a throwable as one of the
arguments. The log formatter would just concat the throwable to
the error message without any whitespace separation.

Some log methods work around this by appending ": " to the error
message, but this is something best resolved in the formatter
itself.