Make unprivileged username a constant, and set it to "anonymous"
I noticed that one place used "unknownUser" while other placing use
"unknown". This commit changes it to "anonymous", which is used by
the studio, and IMO has better semantics. For those that care:

- "anonymous" is known from FTP fame
- "unknown" gives the impression that the client actually has a username
but that Yamcs wasn't able to resolve it. That's a wrong impression.
What it should mean is that the server just doesn't identify users,
therefore everybody is "anonymous".

At any rate, we should probably also include an extra flag to prevent
the case where the user actually has a principal named "anonymous".