Add transitional property defaultInstance to yamcs.yaml
This is in support of Yamcs Studio that should have
capability to connect without specifying an instance.

Currently our websocket handler enforces an instance,
so exposing a default instance to be interpreted by
the client is a decent workaround.

Anecdotal: Would have preferred to called the protobuf
property 'defaultInstance' instead of 'yamcsDefaultInstance'
but that caused compilation errors in proto generated
files, due to overlap with the always-added static methods.