make-client-package.sh: fix issue with sources jar
When there are multiple yamcs-core/yamcs*.jar files,
make-client-package.sh takes the latest one.

But if the latest one is a -sources jar, that one will be chosen
instead of the normal compiled jar file, which will result in a
ClassDefNotFoundException when starting the software because the
.class files are not in the -sources jar.

This commit fixes the issue by excluding the sources jar.