CcsdsPacket: improve toString() and add unit test
The old toString() of CcsdsPacket did not output the final
ASCII bytes of a CcsdsPacket if the size was not a multiple of 16.

It also did not have any unit tests.

This commit improves the toString() by ensuring all ASCII bytes are
output and adds 2 unit tests to test the 16-byte-multiple and the
non-16-byte-multiple cases.