Fix disabled "Apply Selection" button for PP's
On the "Processed Parameters" panel of the Archive Browser,
the "Apply Selection" button is initially disabled and gets
enabled when a date range is selected.

But this behavior was broken in commit 2cbce606244 and the
button did not get enabled anymore, unless you went
to the Telemetry panel and selected a date range first.

This commit restores the button's functionality.