Remove protostuff dependency in favor of protobuf-java-util
protobuf-java-util uses gson to parse or print JSON from protobuf
messages. All occurrences of jackson were removed from the code and
replaced with gson instead.

protobuf-java-util is added as a new dependency to yamcs-core.
Unlike protostuff, it is not added to yamcs-api to avoid a
transitive guava dependency in yamcs-api. Gson does become a new
dependency of yamcs-api (substituting protostuff and jackson)

Note that for the time being our messages are still generated with
protoc 2.6.x. The v3.x.x dependency of protobuf-java can work
with this just fine. It should not be an issue to upgrade to v3 though
since new protoc releases support both v2 and v3 language modes.

Dealing with message extensions was the most problematic part of this
refactoring. Extensions are a proto2 feature, and while they seem
to print just fine via protobuf-java-util, it's not possible to parse
them. Going forward we should probably investigate into migrating our
proto files to v3 as well, then we can migrate to use Any type instead.