Support Sec-WebSocket-Protocol header to negotiate json or protobuf
Previously we were waiting on the first incoming frame, and if it was
binary we'd choose protobuf, else json.

The proper way is to use Sec-WebSocket-Protocol instead, and to default
to JSON if not set (so that browser works without specifying).

While clients are being migrated to this new negotation method, we shall
not yet enforce the default of JSON.