Allow aggregates over infinite windows
In order to implement select count(*) from table, we allow aggregates to run without a window size specification - that is they run until the input stream is closed and then they produce a result.

Using this new infinite windows, implemented the count aggregate so one can now count the records in a table.
To be used with caution as it always makes a full table scan (because in rocksdb there is no way to know how many records in a range).