implemented DB arrays
The yarch list data type has been renamed to array to match SQL standard (in tuples it corresponds  to a java List<>).
Will be used to store the tags associated to a timeline item.

Cleanup exception handling in table serializers: instead of throwing IOException when there is no IO, throw YarchException for unexpected exceptions or DatabaseCorruptionException if data read from the database is inconsistent.