changed the quicksort in IntArray
On profiling it turned out that a lot of time is spent in BasicParameterList.sort because the old implementation of the IntArray sort was doing its worst when the array was already sorted. Because the parameter ids are allocated in the order of their apperance in the packet, and the packet order rarely changes, it followed that the list of parameters was almost always sorted causing the quicksort to always perform at its worst.
So changed to use a pivot in the middle of the list rather than at the end. The current implementation is still 2 times slower than the java implementation (which we cannot use because we need to update two lists in parallel) but the java implementation is much more complicated.