Added XTCE command verifiers and refactor PRM
The PRM has been refactored such as to make it clear that at each delivery (e.g. when a packet is received) there is a list of "current" parameters and the algorithms have to use that in priority and to add the results to the same list (such that dependent algorithms can use the same list in turn).

Only after the list has been updated by algorithms is then sent to all subscribers, such as command verifiers. The verifiers will use the list in priority and if a parameter is not on the list, a verifier will take a value from the last value cache.

The last value cache updated only after all clients have received their subscription.

In the previous implementation the last value cache was updated in between causing verifiers to be confused about what the current value of a parameter is.

In the refactoring process, made the algorithms receive the full parameter list part of a delivery (e.g. all parameters extracted from a packet) but the list is indexed per parameter so an algorithm can select whichever value it wants without browsing through all list. Previous implementation the PRM would select a list of all input parameters for all algorithms and each algorithm had to browse through all list. Which means that the list now is bigger but is indexed.

A particular case which is not yet working (but the refactoring should make it easier to implement) is when one delivery contains multiple values for the same parameter. Theoretically XTCE allows to reference such value in a verifier or algorithm but it is not yet implemented.