refactor match-criteria
The MatchCriteria have a string value to compare pareameters with. Depending on which parameters is to be compared the string value is interpreted as a number or as a string (with some validation on the value which can take). Before this change, the transformation from string to value was done when reading the MDB. That implied knowing the type of the parameter which will be compared with. Because of that one could not use system parameters whose types are not known when loading the MDB.
This change does not attempt to convert the string to a value until much later when all the types are known.

Together with the match criteria there is also a refactoring of the subcription object used in the TM processing to allow easier access to local match criteria evaluators.