java-expression algorithms
Added algorithms of type "java-expressions" - these are bits of Java code compiled by Yamcs on the fly (unlike the java algorithms whcih require a jar to be installed on the server).
Also fixed #539 the inconsistencies in aggregate values/member values - now if the algorithm is declared as needing an aggregate member value as input, the updateInput in AbstractAlgorithmExecutor will be called with that member value rather than the whole aggregate.