Fix calibration of algorithms output
Up until now, the scripts(javascript and python) used to set the outputparam.value. This was sometimes interpreted as a rawValue creating confusion.
From now on, the outputparam.value will always be interpreted as engineering value and the outputparam.rawValue can be used to set the raw value. If the raw value is set but not the engineering value, the engineering value will be derived based on the parameter type (e.g. calibration may be performed)