refactor (unresolved) references
When reading a XTCE or excel file which contains name references to entities (parameter types, containers, etc) defined externally, a bunch of classes xyzReference classes are used. Actions associated to those classes are called when the reference is resolved. For example an action will insert in the parameter definition the the java reference to the parameter type definition.

Before this change, the actions were returning a boolean indicating if the reference was truly resolved. However that was buggy: some actions were added after the reference was already resolved so they could not influence the outcome of the reference solving.

In reviewing the code it turned out that the only reason an action was returning false was because a reference to a parameter was made, the parameter was found but had no type. Things line MatchCriteria need the parameter type to validate that the comparison can be made (e.g. if one tries to compare a string with a number an error occurs). For these references, the actions were checking the type and if it was not there, it was returning false. The code would then first solve the type for the parameter and then call back the action for the parameter reference.

This change eliminates the boolean return to the reference solving actions and enforce the parameter types on all parameters - a parameter reference is not considered solved unless that parameter has a type (anexception is the System Parameters which are created on the fly and they don't have a type until much later).

I also removed a lot of unnecessary duplicate classes, keeping just a NameReference for general references and two specialized ones ParameterReference and ArgumentReference which can solve reference inside aggregate members of parameters respectively arguments.

The only downside I can see in the new implementation is that it breaks a scenario where a parameter depends on its type which depends in turn on the original parameter (with perhaps some other dependencies in between) - that kind of cyclic dependency is anyway not a sane thing to have.