fix the packet names appearing on the front page
the front page shows the TM Processing Statistics - containing some packet names. When a packet matches multiple containers, it is not clear what to show there. For now it shows the name of the container as it is used in the archive - that also makes the  link to the archive behave properly.
In the future the functionality could be extended, maybe in conjunction with having a packet viewer capability on the web.