CCSDS sequence count disconinuity
Do not generate a CCSDS sequence count discontinuity event for the first packet received after Yamcs is restarted.
The message was misleading because it said that the old sequence count was 0, in reality Yamcs (tc preprocessor) didn't know the old sequence.
A better solution would be for Yamcs to remember the sequence counts in between restarts.