Fixed parameter retrieval when only raw values are requested
The parameter samples API with useRawValue=true was causing a NPE. This can be used from graphana. The reason is that the logic of treating raw values equal with engineering values has moved into the SegmentIterator.
Also refactored the parameter info API to use the new ParameterIdDb which provides more capabilities.