bugfix XTCE ref solver
A reference chain Parameter -> Type -> BaseType where BaseType is defined in a different file than Type and the Parameter appears defined before Type resulted in an error because when trying to solve Parameter it did not find Type and it was bailing out. With this change it keeps trying until all references are resolved or no reference is resolved anymore.