Support JDK11-17+, drop JDK 1.8, 9 and 10
Yamcs support of JDK15+ was stalled due to Nashorn
being removed as a built-in to JDK.

The alternative solution on JDK15+ is to use
Nashorn from the classpath, however this was
causing issues with JDK 1.8-10.

Dropping Nashorn is not an option (it is used
for algorithm scripts), so this commit removes
support for JDK 1.8 to 10.

Still, JDK11 has deprecated support for a Nashorn
built-in, so the commit includes a few tricks to
bypass that version.