Added support for Attached Synchro Marker removal
I added a support for Attached Synchro Marker (ASM) removal from a TM frame send to Yamcs through UDP.
ASM is specified in CCSDS 131.0-P-2.1 document (TM SYNCHRONIZATION AND CHANNEL CODING) and its hex value is usually 1ACFFC1D.
I added an asmPresent option in the dataLink part of the yamcs.my_instance.yaml. 
Its value can be either true or false, and added a java case in UdpTmFrameLink.java, removing it if it is said to be present and the frame indeed starts with the ASM. 
Beware, I would have liked to add this in CcsdsFrameDecoder.java but it was a bit harder. So if we do a tcp frame connection, this change is not taken into account. I plan to add this in the future if needed.