Moved message encoding to netty thread
Previously when sending a websocket message the encoding to protobuf binary or json was done on the sender thread (e.g. on the TM link thread if the message was the result of a parameter or packet subscription). Now the message encoding is performed in the netty pipeline on one of the netty http threads.

Incidentally this seems to make the test for #691 not fail anymore even without using the newly implemented low priority option. My guess is that netty picks up the parameter message, encodes it to json and immediately sends it to the network without storing it in the intermediate buffer (it can do that exactly because it runs in the netty i/o thread).