Add minTime, maxTime to sample aggregates #442
This commit also transitions the datatype from the
`time` field from string to google.protobuf.Timestamp
to be consistent with the new fields, as well as the
rest of the API.

This is done in a way that it is invisible to
both protobuf (same serialization id) and JSON
(same name) clients. The new deprecated field
`timeString` will eventually be deleted when
protobuf clients (studio+python) have updated.