fixed TAI offset
When UTC in its actual form has been introduced, there were 10 seconds difference vs TAI. The code was not dealing properly with that causing UTC-TAI difference of 28 instead of 37 seconds.
We now distribute the 9 extra leap seconds in the 9 days before Jan-1972. It is a bit artificial but since UTC was anyway not defined in its actual form before 1972, we deem it adequate.