Load native rocksdbjni from lib path, for distributions
If RocksDB can't find "rocksdbjni" from the system
library path, it tries to find it in the classpath.

This works, but creates a temporary file. When
packing our default distributions, we can avoid this
step by placing the rocksdbjni native library
directly in lib, which is already added to
java.library.path by the yamcsd wrapper.