yamcs cascading
The Yamcs cascading allows one Yamcs server to connect as a client to another (upstream) Yamcs server.
It is realized by a Yamcs Link which provides the following data (each is a sub-link):
- TM packet reception in realtime and archive
- Parameter reception in realtime
- Event reception in realtime
- Command sending and Command History provision

The link is configured with one entry in the links section of the yamcs.<instance>.conf configuration file.