Move towards linkInfo subscription without change detection
This adds a new LinkEvent UPDATE_ALL which will be sent
in parallel to the previous event types.

The UPDATE_ALL event will be sent each 1Hz with info
on all links. Clients should migrate to this new event type,
as the previous event types are now deprecated (desire
to remove the change-detection mechanism).