fix racecondition in replication
It has been noticed that if the thread that is writing to the replication file (which often would be a thread coming from a link) is interrupted in the ReplicationFile#doForceWrite, the replication file gets closed and the ClosedByInterruptException is thrown. This exception would cause the replication to be completely broken from that point on.

This patch will catch that exception (and possibly others) and will forcefully start a new replication file.