Remove few old deprecated functions
This removes the source field from
a command (was replaced a long time
ago by an explicit "assignments" field
(Yamcs has no internal notion of command
"source").

Also removed is the transient UUID which
was associated with issued commands and
that could be used to manipulate queue
entries. This was replaced a while back
with a command id.