fix for array/aggregates without raw value
If an array/aggregate did not have a raw value, the code was erroneusly assuming there was one causing a NPE on retrieval from the cache.
Retrieving from the Parameter Archive was also erroneus because we have an optimization to not store the raw value if it's equal to the engineering value. So because it thought there is an raw value but there was nothing stored in the parameter archive, it was duplicating the engineering into the raw when retrieving from the Parameter Archive.

Fixes #764