Update directory directly #831
This commit starts the next phase in migrating
the storage of user info away from the
"ProtobufDatabase" to the "Yamcs DB".

The ProtobufDatabase is now read only once on
Yamcs startup, then fully migrated to the
directory. Previously (as a precaution) this
operation was done on each change of the user
model.

For a while, for backwards compatibility reasons,
we also mirror the Yamcs DB to the ProtobufDB.

This commit should address a performance issue
that was addressed in #831 when the number
of users scales up by significantly reducing
the amount of queries.