GSS-269: Implement Service-13 as a FileTransferService
This service will internally dispatch telecommands. Some compromises may have to be made with this apporach.
The reason for sending telecommands internally is keep track of them in the cmdHist table, as they might needed for later inspection

At the moment, the mechanism is a fire-and-forget type service. Further features can be added as necessary