fixed memory leak
When compiling an expression a new name was used. That name leaked into the parallelLockMap of the main class loader and never got removed.
The janino SimpleCompiler is using a different class loader for each compilation but that one was querying the parent (which is the main class loader) for a class with that name and that's how the entry in that map was being created.

The code that queries the parent is in the java.lang.classLoader#loadClass:

    protected Class<?> loadClass(String name, boolean resolve)
        throws ClassNotFoundException
    {
        synchronized (getClassLoadingLock(name)) {
            // First, check if the class has already been loaded
            Class<?> c = findLoadedClass(name);
            if (c == null) {
                long t0 = System.nanoTime();
                try {
                    if (parent != null) {
                        c = parent.loadClass(name, false);
                    } else {
                        c = findBootstrapClassOrNull(name);
                    }
                } catch (ClassNotFoundException e) {
                    // ClassNotFoundException thrown if class not found
                    // from the non-null parent class loader
                }

                if (c == null) {
                    // If still not found, then invoke findClass in order
                    // to find the class.
                    long t1 = System.nanoTime();
                    c = findClass(name);

                    // this is the defining class loader; record the stats
                    PerfCounter.getParentDelegationTime().addTime(t1 - t0);
                    PerfCounter.getFindClassTime().addElapsedTimeFrom(t1);
                    PerfCounter.getFindClasses().increment();
                }
            }
            if (resolve) {
                resolveClass(c);
            }
            return c;
        }
    }

The findClass method implemented by janino's ByteArrayClassLoader is eventually called to load the class but at that point parent.loadClass has already been called and the parallelLockMap is poluted.

Since the classloader is always different, the class name does not not need to be different, so this change simply sets the class name for the generated expressions to a fixed value "Expression_Generated".

The problem seems similar with this:
https://issues.apache.org/jira/browse/GROOVY-6655