Allow sparse groups in Parameter Archive
The parameter archive saves data in parameter groups - a parameter group is similar with a table in a relational database.
The parameters that are part of a group are those deliveried together and have the same timestamp - for example those that are extracted from one packet together with ouput of algorithms computed using as inputs those parameters.
Until now, Yamcs did not allow any gap in the groups - each time a delivery contained a different set of parameters than what has been received up to that point, a new group was being created.
For example if an XTCE include condition would cause some parameter to be only sometime extracted from the packet, that would result in a new group. If there were several of those parameters, then the combination of them being present or not present could quikly grow the number of groups.
Also if an algorithm was added, the new parameter computed by the algorithm would add at the end of the delivery so a new group had to be created.
Finally if an array parameter was archived, for each size of the array a different group had to be created because in the parameter archive p[n] is considered as a different basic parameter for each n.

Having so many parameter groups caused a lot of memory to be consumed during parameter archive buildup - because groups are built in parallel up to the maxSegmentSize values. Also during extraction, if a parameter is part of many groups, then parallel iterators had to be created to extract and merge the data, each of those iterators consumign memory and pickink up data from different parts of the disk (which kind of defeated the design of the parameter archive which is supposed to keep values for the same parameter together).

With the current  change sparse groups are allowed - that is not all deliveries that are part of the same group have to have all parameters. The way it works is that with each new delivery, the list of parameters is compared with the existing groups. If the overlap is greater than 0.5 (configurable) then the delivery is considered part of the same group and the group is ammended to include the new parameters. The overlap is defined as the number of parameters common to the two parameter lists divided by the size of the smallest list. That means if one list is included in the other, the overlap is 1. If two existing groups overlap with a parameter delivery by the same amount, then the one with smallest number of parameters is preferred - this is done in order to decrease the number of gaps.

Two options have been added to the parameter archive:
sparseGroups: boolean

If this is true then the sparse group functionality is enabled. If it is false then the parameter archive works like before.

minimumGroupOverlap: double between 0 and 1
The minimum overlap (see above for how it is computed) considered when deciding if one parameter list is part of a group or not.